requirements.txt
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
peewee==3.15.4
pycparser==2.21
PyMySQL==1.0.2
python-decouple==3.6
python-dotenv==0.21.0
Werkzeug==2.2.2
